Broke all components into separate files and basic refactoring
