+++ decoded_files/test_decoded.txt
@@ -1,2 +1,12 @@
-this is a test file.  if you have found this, then you have found the test file. congratulations!  you get to see how i am testing grammar... new lines... and
-all those cr@zy character$ that mäkēs the english language so unique.  i think that i will need more lines... so, here are some of my favorite bands: - metallica - rise against - pink floyd - trampled by turtles - led zeppelin  rock on you guys... shine on you crazy diamonds. 
+this is a test file.  if you have found this, then you have found the test file.
+congratulations!  you get to see how i am testing grammar... new lines... and
+all those cr@zy character$ that mäkēs the english language so unique.
+
+i think that i will need more lines... so, here are some of my favorite bands:
+- metallica
+- rise against
+- pink floyd
+- trampled by turtles
+- led zeppelin
+
+rock on you guys... shine on you crazy diamonds.
