+++ encoded_files/test_output0.txt
@@ -1,1 +1,6 @@
-1110000000101110001010100010100000000000000000000010001010111000111000111011101010111011100010001110111011100010111000101011101000111010100010001010100010100000001110111010001010111000101000!00011101110111000100011101110111000100010111011100010111000100011101000101110100010111000101110101110101110001011101110001110100010111010111010111000111010001011101010001010101000000011101110101000101010100011101011101000$00010111000ä000000000000001011101010000000111011101000100000001110111010111000000011100011101011100010111000000010111010100010001110111011100010111010100010101000000000000001000100011101110001010111010000000111011101110001000111010000000100010111010100000001010001011100011101000-000111010001011101010000000101110100010111010100011101010100010101110001000000000000001011101110100011101000111011101110001110111011100011101110111000101011100010111010111010111000101010100000001110101110111000111010111010001110101110111000101110001110101
+11100000001011100010101000101000000000000000000000100010101110001110001110111010101110111000100011101110111000101110001010111010001110101000100010101000101
+1110111010001010111000101000!00011101110111000100011101110111000100010111011100010111000100011101000101110100010111000101110101110101110001011101110001110100010111010111010111000111010001011101010001010101000000011101110101000101010100011101011101000$00010111000ä000000000000001011101010000000111011101000100000001110111010111
+1110001110101110001011100000001011101010001000111011101110001011101010001010100000000000000100010001110111000101011101000000011101110111000100011101
+1000101110101
+1010001011100011101000-00011101000101110101
+101110100010111010100011101010100010101110001000000000000001011101110100011101000111011101110001110111011100011101110111000101011100010111010111010111000101010100000001110101110111000111010111010001110101110111000101110001110101
