+++ encoded_files/test_output1.txt
@@ -1,1 +1,3 @@
-1010101000101000000011100010111010100000001110101110111000101010100000001110100010101010000000111010001010111000101010111000111011101110000000000000011100010111010100011101011101000101110100010111010100011101110111000000010101110001110000000000000000000001110111000101010001110111010001011100010111010000000000000010001011101011101011100011101010001011101010001110111011100011101011101000111010111011100010111000111000000011100011101011100011100010001010001011101010001010111000000010101110001010111000000010101010000000111000101110111000000011101010001011101000101000101110101110101110001010100010101010000000000000010000000101011101000101110100000001110101000-000111000101000-0001010100011101110100010101000000011101011100011101110111000-0001011100010001110101110111000101110100010101000101110101000111011101010001000000011101011101000111010001010111000111010111011100010111010111010111000101000111011101110001110111011100010111010000000111011100010101
+1010101000101000000011100010111010100000001110101110111000101010100000001110100010101010000000111010001010111000101010111000111011101110000000000000011100010111010100011101011101000101110100010111010100011101110111000000010101110001110000000000000000000001110111000101010001110111010001011100010111010000000000000010001011101011101011100011101010001011101010001110111011100011101011101000111010111011100010111000111000000011100011101011100011100010001010001011101010001010111000000010101110001010111
+10101010000000111000101110111000000011101010001011101000101000101110101110101110001010100010101010000000000000010000000101011101000101110100000001110101000-000111000101000-0001010100011101110100010101000000011101011100011101110111000-0001011100010001110101110111000101110100010101000101110101000111011101010001
+11101011101000111010001010111000111010111011100010111010111010111000101000111011101110001110111011100010111010000000111011100010101
