+++ encoded_files/test_output2.txt
@@ -1,2 +1,4 @@
 101000101010001110000000100010100011101110111000101110001010111010001110101000101000111000000000000001000101011100011100011100000001000111011101110001011100010111000111010000000000000000000001010100010101010001010000000111000000011101110001011101011101011100011101000101110101000101010000000
-00000001010100010111010000000101110100010001110000000ē00010101010001110100010101000101110001011100010101000111010001000101000101000111000000010100011101000000010001110100010111010111010111000111011101110001000101110001010100000001110111000101110001010001110101010001010100000001011100011101011101000000010001011100011100010111011101000000011101011101110000000111011100011101010000000111000000010001000101110101000000011101011100000000000000101010000000111010001110100010101110001011100011101010001110111011100010111010111010111
+00000001010100010111010000000101110100010001110000000ē00010101010001110100010101000101110001011100010101000111010001000101000101000111000000010100011101000000010001110100010111010111010111000111011101110001000101110001010100000001110111000101110001010001110101010001010100000001011100011101011101000000010001011100011100010111011101000000011101011101110000000111011100011101010000000111
+10001000101110101
+11101011100000000000000101010000000111010001110100010101110001011100011101010001110111011100010111010111010111
