+++ encoded_files/test_output3.txt
@@ -1,1 +1,2 @@
-101010000000100010101110100010111010111010111000101011101000101011100010101011100011101110111000000010101000101010100011101011101110001010101000000011101000101010100010001010111010001011101011101011100011101000111000111000101010001110101110111000111011101000111000100011101110111000000011100010100011101110100011101110001011101011101011100010001010001011101011101011100010111000101110001110001000@0001110101110100010111000101110100010101010001110111000101010001000111011101000101010100011101000111011101000111011101110001010001011101011101011100000001110100010101010001010001011101010001000111011100000001000101110101110101110001110111010101110111000101110100010111010001110111011100011101110111000111010111011100010101011100011100010111000111011101110101010001110111000101110101000101110001011101000000010100000001010001010111010001110101000111000101110111010000000111000101110101000-000111010100010111011101000101000101110100000001110101110111000111011101000101110101110101110001010100010000000000000011101110101000101000111010000000
+101010000000100010101110100010111010111010111000101011101000101011100010101011100011101110111000000010101000101010100011101011101110001010101000000011101000101010100010001010111010001011101011101011100011101000111000111000101010001110101110111000111011101000111000100011101110111000000011100010100011101110100011101110001011101011101011100010001010001011101011101011100010111000101110001110001000@00011101011101000101110001011101000101010100011101110001010100010001110111010001010101000111010001110111010001110111011100010100010111010111010111000000011101000101010100010100010111010100010001110111000000010001011101011101011100011101110101011101110001011101000101110100011101110111000111011101110001110101110111000101010111000111000101110001110111011101010100011101110001011101010001011100010111010000000101
+1010001010111010001110101000111000101110111010000000111000101110101000-00011101010001011101110100010100010111010000000111010111011100011101110100010111010111010111000101010001000000000000001110111010100010100011101
