+++ lib/letter_decoder.rb
@@ -1,0 +1,23 @@
+require "./lib/translator"
+
+module ParaMorse
+
+  class LetterDecoder
+    include Translator
+
+    def decode(morse)
+      if valid_morse?(morse)
+        morse_to_eng_translator[morse]
+      else
+        morse.gsub(/[10]/,"")
+      end
+    end
+
+    def valid_morse?(morse)
+      morse.chars.all? do |character|
+        character == "0" || character == "1"
+      end
+    end
+  end
+
+end
