+++ lib/letter_encoder.rb
@@ -1,0 +1,21 @@
+require "./lib/translator"
+
+module ParaMorse
+
+  class LetterEncoder
+    include Translator
+
+    def encode(letter)
+      if valid_letter?(letter)
+        eng_to_morse_translator[letter.downcase]
+      else
+        letter
+      end
+    end
+
+    def valid_letter?(letter)
+      !eng_to_morse_translator[letter.downcase].nil?
+    end
+  end
+
+end
