+++ lib/queue.rb
@@ -1,0 +1,61 @@
+module ParaMorse
+
+  class Queue
+    attr_reader :queue
+
+    def initialize
+      @queue = Array.new
+    end
+
+    def push(digit)
+      queue << digit
+    end
+
+    def pop(amount = 1)
+      if amount == 1
+        queue.shift
+      else
+        pop_multiple(amount).join
+      end
+    end
+
+    def pop_multiple(amount)
+      popped_elements = Array.new
+      amount.times do
+        popped_elements << queue.shift
+      end
+      popped_elements.reverse
+    end
+
+    def pop_all
+      elements = queue.join
+      clear
+      elements
+    end
+
+    def peek(amount = 1)
+      if amount == 1
+        queue[0]
+      else
+        queue[0..amount - 1]
+      end
+    end
+
+    def tail(amount = 1)
+      if amount == 1
+        queue[-1]
+      else
+        queue.reverse[0..amount - 1]
+      end
+    end
+
+    def count
+      queue.length
+    end
+
+    def clear
+      @queue = Array.new
+    end
+  end
+
+end
