+++ test/decoder_test.rb
@@ -1,4 +1,4 @@
-require "./lib/paramorse"
+require "./lib/decoder"
 require "./test/test_helper"
 
 class TestDecoder < Minitest::Test
