+++ test/file_decoder_test.rb
@@ -1,4 +1,5 @@
-require "./lib/paramorse"
+require "./lib/file_decoder"
+require "./lib/file_encoder"
 require "./test/test_helper"
 
 class TestFileDecoder < Minitest::Test
